1. Moved Embedding FX metatype from `experimental/torch/fx` to torch operator metatypes and initialized a new FX metatype registry. 2. Refactored fx weights compression backend for reusability. 3. Include Fx embedding metatype in Torch weights compression backend for reusability in both backends.
