+++ nncf/experimental/torch/fx/nncf_graph_builder.py
@@ -20,7 +20,6 @@ from nncf.common.graph.layer_attributes import ConvolutionLayerAttributes
 from nncf.common.graph.layer_attributes import Dtype
 from nncf.common.graph.operator_metatypes import UnknownMetatype
 from nncf.common.logging import nncf_logger
-from nncf.experimental.torch.fx import operator_metatypes as fx_om
 from nncf.experimental.torch.fx.node_utils import get_tensor_constant_from_node
 from nncf.torch.dynamic_graph.layer_attributes_handlers import apply_args_defaults
 from nncf.torch.graph.graph import PTNNCFGraph
@@ -77,7 +76,7 @@ class GraphConverter:
         if metatype in [om.PTEmbeddingMetatype]:
             weight_node = node.args[0]
             if weight_node.op == "get_attr":
-                return fx_om.FXEmbeddingMetatype
+                return om.FXEmbeddingMetatype
 
         return metatype
