+++ nncf/torch/graph/operator_metatypes.py
@@ -28,6 +28,7 @@ from nncf.torch.dynamic_graph.structs import NamespaceTarget
 ModuleAttributes = TypeVar("ModuleAttributes", bound=BaseLayerAttributes)
 
 PT_OPERATOR_METATYPES = OperatorMetatypeRegistry("operator_metatypes")
+FX_OPERATOR_METATYPES = OperatorMetatypeRegistry("operator_metatypes")
 
 
 class PTOperatorMetatype(OperatorMetatype):
@@ -916,6 +917,14 @@ class PTEmbeddingMetatype(PTOperatorMetatype):
     hw_config_names = [HWConfigOpName.EMBEDDING]
     subtypes = [PTModuleEmbeddingMetatype]
     weight_port_ids = [1]
+
+
+@FX_OPERATOR_METATYPES.register()
+class FXEmbeddingMetatype(OperatorMetatype):
+    name = "EmbeddingOp"
+    module_to_function_names = {NamespaceTarget.ATEN: ["embedding"]}
+    hw_config_names = [HWConfigOpName.EMBEDDING]
+    weight_port_ids = [0]
 
 
 @PT_OPERATOR_METATYPES.register(is_subtype=True)
